requirements.txt
fastapi[standard]
pymongo
python-dotenv
cloudinary
python-multipart
bcrypt
pyjwt
